+++ nodes/Autotask/operations/common/picklist-labels/description.ts
@@ -11,7 +11,7 @@ export const picklistLabelOption: INodeProperties = {
 	description: 'Whether to automatically add "_label" fields for picklist fields with human-readable values (no labels will be added for empty values)',
 	displayOptions: {
 		show: {
-			operation: ['get', 'getMany', 'getManyAdvanced', 'whoAmI'],
+			operation: ['get', 'getMany', 'getManyAdvanced', 'getUnposted', 'getPosted', 'whoAmI'],
 		},
 	},
 	typeOptions: {